feat: 그리드 계산 유틸 분리
